[hw4: nodejs] добавлена поддержка POST-параметров через body-parser
