+++ README.md
@@ -117,8 +117,8 @@ $ npm run server
 - [x] При передаче get-параметра type включается фильтрация по типам событий. Возможна фильтрация по нескольким типам событий, пока все типы валидны (/api/events?type=info:critical)
 - [x] При передаче некорректного type — отдавать статус 400 "incorrect type"
 - [x] Все остальные роуты должны отдавать `<h1>Page not found</h1>`, с корректным статусом 404
+- [x] Добавлена поддержка POST-параметров
 - [ ] Подключить данные к вёрстке из первого задания
-
 
 ## Задание
 
@@ -132,11 +132,11 @@ $ npm run server
 В одной из карточек реализовать управление камерой умного дома.
 
 
-### ДЗ #3 
+### ДЗ #3 «Мультимедиа»
 [Задание](https://github.com/mad-gooze/shri-2018-2-multimedia-homework)
 
 В этом задании мы будем делать страницу видеонаблюдения в умном доме.
 На странице можно смотреть HLS видеопотоки с веб-камер, переключать камеры, накладывать риал-таймовые фильтры на видео и следить за активностью в умном доме.
 
-### ДЗ #4 
+### ДЗ #4 «Node js»
 Написать сервер на express который будет подниматься на 8000 порту и обрабатывать два роута: /status и /api/events
