+++ package.json
@@ -31,6 +31,7 @@
     "webpack-dev-server": "^3.1.9"
   },
   "dependencies": {
+    "body-parser": "^1.18.3",
     "express": "^4.16.4",
     "hls.js": "^0.11.0"
   }
