+++ server/index.js
@@ -1,46 +1,30 @@
 const express = require('express');
+const bodyParser = require('body-parser');
+const {pad, dateDiff, filterEvents, checkFilter} = require('./utils');
 const eventsJson = require('../events.json');
 
 const app = express();
 const startWorkDate = new Date();
-const TYPES = ['info', 'critical'];
 
-function pad(num, length = 2, symb = '0') {
-  const n = num.toString();
-  const result = n.length >= length ? n : pad(symb + n, length, symb);
-  return result;
-}
 
-function dateDiff(date1, date2) {
-  let diff = date2 - date1;
-
-  let hours = Math.floor(diff / 60 / 60 / 1000);
-  let minutes = Math.floor(diff / 60 / 1000) % 60;
-  let seconds = Math.floor(diff / 1000) % 60;
-
-  const result = `${pad(hours)}:${pad(minutes)}:${pad(seconds)}`;
-  return result;
-}
-
-function filterEvents(query) {
-  if (!query) return JSON.stringify(eventsJson, null, 2);
-  const types = query.split(':');
-
-  const dataEvents = [...eventsJson.events]
-
-  const filtredEvents = dataEvents.filter(item => types.includes(item.type) );
-
-  const result = {
-    events: filtredEvents,
+function eventsView(type, res) {
+  // Если невалидный запрос
+  if (type && !checkFilter(type)) {
+    res.send('incorrect type', 400);
+    return;
   }
 
-  return JSON.stringify(result, null, 2);
+  const result = filterEvents(type, eventsJson);
+
+  res.setHeader('Content-Type', 'application/json');
+  res.send(result);
 }
 
-function checkQuery(type, query) {
-  const queries = query[type].split(':');
-  return !queries.some(item => TYPES.includes(item) === false );
-}
+// parse application/x-www-form-urlencoded
+app.use(bodyParser.urlencoded({ extended: false }))
+
+// parse application/json
+app.use(bodyParser.json())
 
 app.get('/', function (req, res) {
   res.send('Hello World!');
@@ -53,17 +37,13 @@ app.get('/status', function (req, res) {
 });
 
 app.get('/api/events', function (req, res) {
+  const {type} = req.query;
+  eventsView(type, res);
+});
 
-  // Если невалидный запрос
-  if (req.query.type && !checkQuery('type', req.query)) {
-    res.send('incorrect type', 400);
-    return;
-  }
-
-  const result = filterEvents(req.query.type);
-
-  res.setHeader('Content-Type', 'application/json');
-  res.send(result);
+app.post('/api/events', function (req, res) {
+  const {type} = req.body;
+  eventsView(type, res);
 });
 
 app.get('*', function(req, res){
